requirements.txt
altgraph==0.17.4
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
certifi==2025.1.31
et_xmlfile==2.0.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
pyaes==1.6.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
pywin32-ctypes==0.2.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
TgCrypto==1.2.5
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
